refactor(api): update login endpoint paths and add currency examples
- Change login endpoint from `/login` to `/auth/login` in account and currencies request files for consistency.
- Introduce new example tests for currency operations, demonstrating basic functionalities such as creation, addition, and searching of currencies.
- Remove outdated example files to streamline the codebase and improve clarity.
