+++ docs/requests/account.http
@@ -10,7 +10,7 @@ Content-Type: application/json
 
 ### Login
 # @name login
-POST {{host}}/login HTTP/1.1
+POST {{host}}/auth/login HTTP/1.1
 Content-Type: application/json
 
 {
