+++ pkg/registry/currency_example_test.go
